Add "complete" static linking
All system libraries are now also statically linked into the binary
with the exception of 'glibc' libraries.  The 'librdkafka' library
uses 'getaddrinfo' which "requires at runtime the shared libraries
from the glibc version used for linking".  This may be a portability
issue if you build and run on different systems.